refactor(#294): separate artifact router factories
